feat(empower-core): add event handlers to options
This moves the `onError` handler into options, and introduces an `onSuccess` handler.
Sensible defaults for both are provided in `default-options`.

This will be really handy for AVA, since we track both successfull and failed assertions
for comparison against the plan count.

Both handlers are invoked outside of a try/catch block, so either can elect to throw,
and the error will not be intercepted. (It is unlikely anyone would ever want to do
that in `onSuccess`, but it wouldn't be good for us to mask that error either).

I also dropped the `saveContextOnRethrow` and `modifyMessageOnRethrow` options entirely.
They were not being used anywhere in this module. The main `empower` module can wrap this
one and provide custon `onError` and `onSuccess` callbacks that are sensitive to those options.